Interface

TODO: read and fix some referencing issues

The @tensor macro rewrites tensor operations in terms of basic building blocks, such that any tensor type that implements the following interface can be supported. In these methods, C will indicate an output tensor which is changed in-place, while A and B denote input tensors that are left unaltered. pC, pA and pB denote an Index2Tuple, a tuple of two tuples that represents a permutation and partition of the original tensor indices. Finally, conjA and conjB are symbols that are used to indicate if the input tensor should be conjugated (:C) or used as-is (:N).

Operations

Missing docstring.

Missing docstring for tensoradd!(::Any, ::Index2Tuple, ::Any, ::Symbol, ::Number, ::Number). Check Documenter's build log for details.

Missing docstring.

Missing docstring for tensortrace!(::Any, ::Index2Tuple, ::Any, ::Index2Tuple, ::Symbol, ::Number, ::Number). Check Documenter's build log for details.

Missing docstring.

Missing docstring for tensorcontract!(::Any, ::Index2Tuple, ::Any, ::Index2Tuple, ::Symbol, ::Any, ::Index2Tuple, ::Symbol, ::Number, ::Number). Check Documenter's build log for details.

TensorOperations.tensorscalar — Function

tensorscalar(C)

Return the single element of a tensor-like object with zero indices or dimensions as a value of the underlying scalar type.

Allocations

For some networks, it will be necessary to allocate output and/or intermediate tensors. This process is split into the following hierarchy of functions, where custom tensor types can opt in at different levels.

By default, the process is split into three steps.

First, the scalar type TC of the resulting tensor is determined. This is done by leveraging VectorInterface.jl's scalartype, and promoting the results along with the types of any scalars that are present.

Missing docstring.

Missing docstring for TensorOperations.promote_add. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TensorOperations.promote_contract. Check Documenter's build log for details.

Then, the type and structure of the resulting tensor is determined. The former represents all the information that is contained within the type, while the latter adds the required runtime information (e.g. array sizes, ...).

TensorOperations.tensoradd_type — Function

tensoradd_type(TC, pC, A, conjA)

Obtain typeof(C), where C is the result of tensoradd!(C, pC, A, conjA) with scalar type TC.

TensorOperations.tensoradd_structure — Function

tensoradd_structure(pC, A, conjA)

Obtain the structure information of C, where C would be the output of tensoradd!(C, pC, A, conjA).

TensorOperations.tensorcontract_type — Function

tensorcontract_type(TC, pC, A, pA, conjA, B, pB, conjB)

Obtain typeof(C), where C is the result of tensorcontract!(C, pC, A, pA, conjA, B, pB, conjB) with scalar type TC.

TensorOperations.tensorcontract_structure — Function

tensorcontract_structure(pC, A, pA, conjA, B, pB, conjB)

Obtain the structure information of C, where C would be the output of tensorcontract!(C, pC, A, pA, conjA, B, pB, conjB).

Finally, the tensor is allocated, where a flag indicates if this is a temporary object, or one that will persist outside of the scope of the macro. If the resulting tensor is a temporary object and its memory will not be freed by Julia's garbage collector, it can be explicitly freed by implementing tensorfree!, which by default does nothing.

TensorOperations.tensoralloc — Function

tensoralloc(ttype, structure, istemp=false, [backend::Backend])

Allocate memory for a tensor of type ttype and structure structure. The optional third argument can be used to indicate that the result is used as a temporary tensor, for which in some cases and with some backends (the optional foruth argument) a different allocation strategy might be used.

TensorOperations.tensorfree! — Function

tensorfree!(C)

Provide a hint that the allocated memory of C can be released.

The @tensor macro will however only insert the calls to the following functions, which have a default implementation in terms of the functions above.

Missing docstring.

Missing docstring for TensorOperations.tensoralloc_add. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TensorOperations.tensoralloc_contract. Check Documenter's build log for details.

Utility

Some of the optional keywords for @tensor can be accessed only after implementing the following utility functions:

TensorOperations.tensorcost — Function

tensorcost(A, i)

Compute the contraction cost associated with the ith index of a tensor, such that the total cost of a pairwise contraction is found as the product of the costs of all contracted indices and all uncontracted indices.

TensorOperations.checkcontractible — Function

checkcontractible(A, iA, conjA, B, iB, conjB, label)

Verify whether two tensors opA(A) and opB(B) are compatible for having their respective index iA and iB contracted, and throws an error if not. The operation opA acts as identity if conjA equals :N and as conj if conjA equals :C; the operation opB is determined by conjB analogously.